Move endpoints to standalone functions
